1. Updated retain calls to be on HttpContent objects instead of internal ByteBuf 2. Updated release strategy to use SimpleChannelInboundHandler autorealease 3. Updated maxContentLength from type int to long

+++ codec-http2/src/main/java/io/netty/handler/codec/http2/Http2HttpDecoderSettingsNotifier.java
@@ -26,12 +26,12 @@ import java.util.List;
 public class Http2HttpDecoderSettingsNotifier extends Http2HttpDecoder {
   private List<Http2EventListener<Http2Settings>> settingsListeners;
 
-  public Http2HttpDecoderSettingsNotifier(int maxContentLength) {
+  public Http2HttpDecoderSettingsNotifier(long maxContentLength) {
     super(maxContentLength);
     init();
   }
 
-  public Http2HttpDecoderSettingsNotifier(int maxContentLength, boolean validateHttpHeaders) {
+  public Http2HttpDecoderSettingsNotifier(long maxContentLength, boolean validateHttpHeaders) {
     super(maxContentLength, validateHttpHeaders);
     init();
   }
